refactor composite trip creation into composite_trip_creation.py
