+++ emission/tests/common.py
@@ -2,6 +2,7 @@ from __future__ import print_function
 from __future__ import unicode_literals
 from __future__ import division
 from __future__ import absolute_import
+import emission.analysis.plotting.composite_trip_creation
 # Standard imports
 from future import standard_library
 standard_library.install_aliases()
@@ -195,7 +196,7 @@ def runIntakePipeline(uuid):
     eacilp.infer_labels(uuid)
     eaue.populate_expectations(uuid)
     eaum.create_confirmed_objects(uuid)
-    eaum.create_composite_objects(uuid)
+    emission.analysis.plotting.composite_trip_creation.create_composite_objects(uuid)
 
 def configLogging():
     """
